Add files via upload
Data logging receptionist
It is the day of interview in a company and the chance is only for first 10 candidates. The receptionist has the duty of noting the name and degree of every candidate. She was using a register book so far but wishes to switch this task to a computer operation. So, she requires your Python programming skills to make a solution for her problem.