Update CoffeeRating.ipynb
Coffee rating
'Aroma' Coffee factory has started producing a new blend of coffee. They want to know the customer feedback for this new variety that they decided to give people for a try on this coffee. On this regard, they setup a coffee kiosk in front of INOX movie theatres intended to deliver a free cup of coffee to outgoing viewers after the movie. The only thing people have to do is to remark their rating after the coffee in a special alphanumeric feedback device available at the kiosk. The rating is accepted in a scale of 1-5 (Terrible - Tongueblaster). When all people completes giving feedback, the staff enters a password to see the average rating for the coffee.

NOTE:
1. Logically, it is not possible to get an initial count of number of people who are willing to drink the coffee
2. Some people might leave after having the coffee without marking the feedback
3. Rating value is not considered other than 1-5